requirements.txt
pandas>=1.0.0
numpy>=1.0.0
PyYAML>=5.1
matplotlib>=3.0.0
scikit-learn>=0.24.0
ypstruct
PyYAML